% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/d3graph.R
\name{d3.graph}
\alias{d3.graph}
\title{d3.graph}
\usage{
d3.graph(df, nodeMinSize = 6, nodeMaxSize = 12, edgeMinStroke = 0.05,
  edgeMaxStroke = 0.25, circleFillOpacity = 1, circleStroke = "black",
  circleStrokeWidth = 1, colors = c("#FF475C", "#F65C44", "#EE8442",
  "#E5A940", "#DDCB3D", "#BFD53B", "#93CC39", "#69C436", "#43BB34", "#32B344",
  "#30AA5F"), width = NULL, height = NULL)
}
\arguments{
\item{df}{A data.frame which contains graph data. Some columns are expected. Every row is an edge. \cr
\itemize{
\item{"source"}{Required column. ID or Label of source/from node of the edge.}
\item{"target"}{Required column. ID or Label of target/to node of the edge.}
\item{"sourceLabel"}{Optional column. Label of source/from node of the edge}
\item{"targetLabel"}{Optional column. Label of target/to node of the edge}
\item{"sourceSize"}{Optional column. Size of source node of the edge. ize about 5-10 plots better.}
\item{"targetSize"}{Optional column. Size of target node of the edge. ize about 5-10 plots better.}
\item{"weight"}{Optional column. Weight of the edge.}
}}

\item{circleFillOpacity}{Opacity of nodes.}

\item{circleStroke}{Color of stroke of nodes. In any format an html can understand.}

\item{circleStrokeWidth}{Size of stroke of nodes.}

\item{colors}{Array of colors in any format an html can understand.}

\item{width}{Width of svg}

\item{height}{Height of svg}
}
\value{
An htmlwidget.
}
\description{
\code{d3.graph} creates and htmlwidgets using force networks in D3. Inputs must be a data.frame in some specific format.
}
\details{
Inputs must be a data.frame, not data.table. It is expected to fix this in the future.
}

