% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/d3graphsankey.R
\name{d3.graphSankeyOutput}
\alias{d3.graphSankeyOutput}
\title{Widget output function for use in Shiny}
\usage{
d3.graphSankeyOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

