% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/d3radarplot.R
\name{d3.radarplotOutput}
\alias{d3.radarplotOutput}
\title{Widget output function for use in Shiny}
\usage{
d3.radarplotOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

