% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/d3radarplot.R
\name{renderD3.radarplot}
\alias{renderD3.radarplot}
\title{Widget render function for use in Shiny}
\usage{
renderD3.radarplot(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

